renderer: float-precision cluster, WIP UNBENCHED (Refs #17)
Net settled state of upstream 30fdd88c + 59b1262b + 6f0736ce
(float precision loss in renderer): shaderTime/floatTime/clampTime/
timeOffset widened to double; WAVEVALUE, turbulent, bulge-off and
animated-image index use int64_t; R_NoiseGet4f t param -> double;
R_BindAnimatedImage 64-bit modulus -> subtract loop.

Compiles clean on ppc750/ppc7400/x86_64/i386. NOT yet benched on
G3/G4 - int64_t casts sit in the wave/turbulent hot path, so this
needs a hardware bench pass and may be reverted as a recorded
negative. Parked on a branch, deliberately NOT on master, so it
stays out of any release build until verified.

+++ code/renderer/tr_local.h
@@ -396,8 +396,8 @@ typedef struct shader_s {
 
 	void		(*optimalStageIteratorFunc)( void );
 
-  float clampTime;                                  // time this shader is clamped to
-  float timeOffset;                                 // current time offset for this shader
+  double clampTime;                                 // time this shader is clamped to
+  double timeOffset;                                // current time offset for this shader
 
   int numStates;                                    // if non-zero this is a state shader
   struct shader_s *currentShader;                   // current state if this is a state shader
@@ -438,7 +438,7 @@ typedef struct {
 	byte		areamask[MAX_MAP_AREA_BYTES];
 	qboolean	areamaskModified;	// qtrue if areamask changed since last scene
 
-	float		floatTime;			// tr.refdef.time / 1000.0
+	double		floatTime;			// tr.refdef.time / 1000.0
 
 	// text messages for deform text shaders
 	char		text[MAX_RENDER_STRINGS][MAX_RENDER_STRING_LENGTH];
@@ -1150,7 +1150,7 @@ extern cvar_t	*r_marksOnTriangleMeshes;
 
 //====================================================================
 
-float R_NoiseGet4f( float x, float y, float z, float t );
+float R_NoiseGet4f( float x, float y, float z, double t );
 void  R_NoiseInit( void );
 
 void R_SwapBuffers( int );
@@ -1346,7 +1346,7 @@ typedef struct shaderCommands_s
 	color4ub_t	constantColor255[SHADER_MAX_VERTEXES] QALIGN(16);
 
 	shader_t	*shader;
-	float		shaderTime;
+	double		shaderTime;
 	int			fogNum;
 
 	int			dlightBits;	// or together of all vertexDlightBits
